MPU Viewer Tool done

+++ src/gui/text_utils.h
@@ -8,5 +8,5 @@ std::string toHexStr(uint16_t number);
 bool isValidHex8(const std::string& str);
 bool isValidHex16(const std::string& str);
 
-uint8_t fromHexStr8(const std::string& str);
-uint8_t fromHexStr16(const std::string& str);
+uint8_t  fromHexStr8(const std::string& str);
+uint16_t fromHexStr16(const std::string& str);
